Phase 4: extract useAppNavigation from App.tsx
Everything that reads or writes the URL was scattered across App.tsx:
the route state at the top, the popstate/hashchange listener 180 lines
down, then navigate, exitOverlay and five nav handlers another 350 lines
past that. They are one thing — the app's router — and none of them is
about rendering a route.

- useAppNavigation.ts  route state and its sync effect, navigate,
  exitOverlay + playReturnPathRef, and handleCreateNav /
  handlePartyCreateNav / handlePartyNav / handleConnectNav /
  handleHomeAnchorNav

The scoping plan left `route` in App and passed `setRoute` down. Reading
the file, setRoute has exactly three writers — the initial state, the
sync effect, and navigate — and all three move here, so App would have
been holding a setter purely to hand it back. The hook owns the state and
returns `route` instead.

Two handlers reach outside the router: handlePartyCreateNav seeds the
hero prompt and handleHomeAnchorNav queues a scroll target. Those stay
App's state, taken as parameters.

Handlers stay plain functions rather than becoming useCallback — they are
rebuilt every render today, inside an object literal that is rebuilt
anyway. App.tsx 1212 -> 1112.
